+++ app/controllers/welcome_controller.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
 class WelcomeController < ApplicationController
-  def index; end
+  def index
+    @docs = Doc.all
+  end
 end
