+++ app/views/administrate/docs/_form.html.erb
@@ -1,4 +1,4 @@
-<%= simple_form_for @doc do |f| %>
+<%= simple_form_for [:administratre, @doc] do |f| %>
   <%= f.input :title %>
   <%= f.input :content %>
   <%= f.button :submit, class: "button" %>
