+++ app/views/administrate/docs/edit.html.erb
@@ -2,5 +2,5 @@
   <h1>Fix Doc!</h1>
   <%= render 'form' %>
 
-  <%= link_to "Cancel", doc_path %>
+  <%= link_to "Cancel", administrate_doc_path %>
 </div>
