+++ app/views/administrate/docs/index.html.erb
@@ -13,7 +13,7 @@
     <% else %>
       <h2>Create Doc!</h2>
       <p>Start Creating Document and Organizing your life!</p>
-      <%= button_to "Create Doc", new_doc_path, method: :get, class: "button" %>
+      <%= button_to "Create Doc", new_administrate_doc_path, method: :get, class: "button" %>
     <% end %>
   </div>
 </div>
