+++ app/views/administrate/docs/new.html.erb
