+++ app/views/welcome/_doc.html.erb
@@ -1,0 +1,6 @@
+<div class="">
+  <p>
+    <strong>Title:</strong>
+    <%= doc.title%>
+  </p>
+</div>
