+++ app/views/welcome/index.html.erb
@@ -1,26 +1,7 @@
-<!-- Banner -->
-<div id="banner">
-  <div class="banner_content">
-    <h1>Hub</h1>
-    <p>Your life's Work. Organized, always.</p>
-    <%= link_to "Enter", new_user_session_path %>
-    <%= link_to "Join Now", new_user_registration_path, class: "button" %>
-  </div>
-</div>
+<h1>Docs</h1>
 
-<!-- Callouts -->
-<div class="callout_inner">
-  <div class="wrapper">
-    <div class="callout">
-      <h2>Organize</h2>
-      <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Repellat autem nulla culpa, officiis in saepe natus placeat quaerat voluptatum cumque at exercitationem fugit corrupti eius sapiente molestias provident nam perferendis?</p>
-    </div>
-  </div>
-</div>
-
-<!-- Bottom CTA -->
-<div class="wrapper">
-  <h2>Go Ahead!</h2>
-  <br>
-  <%= link_to "Join Now", new_user_registration_path, class: "button" %>
+<div id="docs">
+  <% @docs.each do |doc| %>
+  <%= render partial: "doc", locals: {doc: doc} %>
+  <% end %>
 </div>
