+++ config/routes.rb
@@ -2,12 +2,12 @@
 
 Rails.application.routes.draw do
   devise_for :users
-  get "welcome/index"
-
-  resources :docs
-  authenticated :user do
-    root "docs#index", as: "authenticated_root"
+  namespace :administrate do
+    resources :docs
   end
+
+  #get "welcome/index"
+  #root "docs#index", as: "authenticated_root"
 
   root "welcome#index"
 end
